+++ BetterCombat/Patches/Vanilla/CombatManeuvers/ImprovedCombatManeuverFeats.cs
@@ -6,10 +6,7 @@ using Kingmaker.Blueprints.Classes;
 using Kingmaker.RuleSystem.Rules;
 using Kingmaker.UnitLogic.FactLogic;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BetterCombat.Patches.Vanilla.CombatManeuvers
 {