+++ BetterCombat/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.2.1
+
+### Fixed
+- Combat-Maneuver-as-attack-toggles are now a bit smarter and won't use "redundant" Combat Maneuvers (e.g. tripping a target that is already (going) prone).
+- Tande Trip no longer has any prerequisites
+
 ## 1.2.0
 
 ### Added